Add support for fraction of minute

+++ migrations/225_assessment_questions__grade_rate__add.sql
@@ -1,1 +1,1 @@
-ALTER TABLE assessment_questions ADD COLUMN grade_rate_minutes integer default 0;
+ALTER TABLE assessment_questions ADD COLUMN grade_rate_minutes double precision default 0;
